Analyse du code

+++ app/src/main/java/com/openclassrooms/tajmahal/ui/restaurant/DetailsFragment.java
@@ -178,7 +178,7 @@ public class DetailsFragment extends Fragment {
 
         binding.tvReviewsAverage.setText(sAvg);
 
-        float fNote = (float) (rAverage.floatValue());
+        float fNote = rAverage.floatValue();
         binding.rbReviewsAverage.setRating(fNote);
 
     }
